review: state invariants without ticket history, add the named-colleague negative case
Fable and Codex both flagged the same issues in the first pass:

- Several new comments cited issue numbers and "before this fix"/"what this
  actually needed" narration — AGENTS.md's rule against build-process residue
  in comments. Rewritten to state the enduring invariant on its own.
- open-deals-per-company's reverted classification left a stale comment
  calling it "install-wide" right beside the correct one two tests later.
  Split stage-age and open-deals-per-company into their own files instead of
  leaving them in report_dealsbystage_integration_test.go, where neither
  belongs.
- Two pre-existing comments this diff made false, left uncorrected:
  reportpopulation.go's "the ad-hoc datasource plan is the one surface that
  takes it" (now six more specs do), and a fixture comment in
  analyticsmcpquestions_manager_integration_test.go claiming every report in
  that file measures the caller's own population (win-loss doesn't anymore).
- No test pinned the accountability boundary itself: that naming a colleague
  or team explicitly stays exact even though the caller's own default now
  admits an unowned row. Added
  TestATypedQueryNamingAColleagueExcludesAnUnownedRow through the real
  typed-query door a caller uses to name someone else — confirmed it fails
  (6 instead of 5) with the self-check guard bypassed, and passes with it
  restored.
- An unchecked type assertion in the new leads-by-status test now matches its
  sibling test's checked form.
