+++ backend/internal/compose/analyticsquery_integration_test.go
@@ -493,6 +493,40 @@ func TestATypedQueryRefusesAPopulationTheAskerCannotMeasure(t *testing.T) {
 	}
 }
 
+// Naming a colleague explicitly is a different ask than measuring your own
+// default population, and an unowned row must not count toward it: this is a
+// standing forecast asking what a NAMED person specifically committed to,
+// and an unclaimed deal is nobody's commitment. Only the caller's own
+// default (self, or a team manager's own managed-teams default) admits an
+// unowned row — never an explicitly named owner or team.
+func TestATypedQueryNamingAColleagueExcludesAnUnownedRow(t *testing.T) {
+	e := setupForecast(t)
+	amount := int64(100_000)
+	// Five of Rep3's own, clearing analyticsquery.DefaultFloor, plus one
+	// unowned deal that must not be counted as Rep3's.
+	for i := 0; i < 5; i++ {
+		e.seedOpenDeal(t, "Rep3's own", 20, &e.Rep3, &amount, nil)
+	}
+	e.seedOpenDeal(t, "Unowned", 20, nil, &amount, nil)
+
+	answer, err := e.askAnalytics(e.wideLensCtx(e.Rep1), t, analyticsquery.Query{
+		Entity:    "open-deals-per-company",
+		ScopeKind: ScopeKindOwner,
+		ScopeID:   e.Rep3.String(),
+		Measures:  []analyticsquery.Measure{{Fn: analyticsquery.CountAll, As: "deals"}},
+	})
+	if err != nil {
+		t.Fatalf("naming a colleague by id was refused: %v", err)
+	}
+	if len(answer.Rows) != 1 {
+		t.Fatalf("an ungrouped query answered %d rows: %+v", len(answer.Rows), answer.Rows)
+	}
+	if got := answer.Rows[0]["deals"]; got != float64(5) && got != int64(5) {
+		t.Errorf("Rep3's named count is %v, want exactly their own 5 — the unowned "+
+			"deal must not be counted as a named colleague's commitment", got)
+	}
+}
+
 // wideLensCtx is a REAL seat that may measure the whole workspace.
 //
 // A real one because saving a run records who asked, and report_run carries a
